insight_test/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pandas==2.2.1
python-dotenv==1.0.1
anthropic==0.3.11
httpx==0.24.1  # Required by anthropic
